added arguments for the bash script, fixed some tab-space issues and …

+++ dspace_metadata_import.sh
@@ -3,7 +3,7 @@
 
 ########################################################################################
 ## Run this script to import metadata from old dspace into new dspace6 database, and  ##
-## copy over the metadata registry files.											  ##
+## copy over the metadata registry files.                                             ##
 ########################################################################################
 
 
@@ -11,29 +11,72 @@
 ## DSpace6 Upgrading: https://wiki.duraspace.org/display/DSDOC6x/Upgrading+DSpace ##
 
 
-#########################################################################
-#                       									 			#
-#	Need to 															#
-#		1) copy over the dumped pgsql file from old dspace, and			#
-#		2) set up the ownership, and									#
-#		3) replace the user to be dspace or something, and 				#
-#		4) comment out the block of adding rule protect_approval_upd	#
-#    	5) comment out the privileges granted to user libacct			#
-#                       												#
-#########################################################################
+#################################################################################
+#                                                                               #
+#       Need to                                                                 #
+#		1) copy over the dumped pgsql file from old dspace, and         #
+#		2) set up the ownership, and                                    #
+#		3) replace the user to be dspace or something, and              #
+#		4) comment out the block of adding rule protect_approval_upd    #
+#               5) comment out the privileges granted to user libacct           #
+#                                                                               #
+################################################################################
 
-PATH_TO_REGISTRIES_FROM_OLD_DSPACE=/tmp/registries.tar.gz
-PATH_TO_OLD_DSPACE_DATABASE_DATA=/tmp/test3.pgsql
-EXISTING_DSPACE_DATABASE_NAME=dspace
-EXISTING_DSPACE_DATABASE_NAME_CHANGED=dspace_old
-NEW_DSPACE_DATABASE_NAME=dspace2
-DSPACE_DATABASE_OWNER=dspace
+
+EXISTING_DSPACE_DATABASE_NAME_CHANGED=dspace_changed_from_existing_temp
+NEW_DSPACE_DATABASE_NAME=dspace_new_temp
+
+# while getopts old_registries_file_path:old_dspace_data_file_path:existing_dspace_database_name:dspace_owner_name:database_homename: option
+#   do
+#     case "${option}"
+#     in
+#       old_registries_file_path)       PATH_TO_REGISTRIES_FROM_OLD_DSPACE=${OPTARG};;
+#       old_dspace_data_file_path)      PATH_TO_OLD_DSPACE_DATABASE_DATA=${OPTARG};;
+#       existing_dspace_database_name)  EXISTING_DSPACE_DATABASE_NAME=${OPTARG};;
+#       dspace_owner_name)              DSPACE_DATABASE_OWNER=${OPTARG};;
+#       database_homename)              DATABASE_HOSTNAME=$OPTARG;;
+    
+#  esac
+# done
+
+PATH_TO_REGISTRIES_FROM_OLD_DSPACE=$1
+PATH_TO_OLD_DSPACE_DATABASE_DATA=$2
+EXISTING_DSPACE_DATABASE_NAME=$3
+DSPACE_DATABASE_OWNER=$4
+DATABASE_HOSTNAME=$5
+
+
+if [[ -z "${PATH_TO_REGISTRIES_FROM_OLD_DSPACE// }" ]]; then
+	echo "ERROR: the path to the registries of the old dspace instance is not defined!"
+	exit
+fi
+
+if [[ -z "${PATH_TO_OLD_DSPACE_DATABASE_DATA// }" ]]; then
+	echo "ERROR: the path to the database data file of the old dspace instance is not defined!"
+	exit
+fi
+
+if [[ -z "${EXISTING_DSPACE_DATABASE_NAME// }" ]]; then
+	echo "ERROR: the name of the database of the existing dspace instance is not defined!"
+	exit
+fi
+
+if [[ -z "${DSPACE_DATABASE_OWNER// }" ]]; then
+	echo "ERROR: the name of the dspace database owner is not defined!"
+	exit
+fi
+
+if [[ -z "${DATABASE_HOSTNAME// }" ]]; then
+	echo "ERROR: the hostname of the dspace database is not defined!"
+	exit
+fi
+
 
 sudo systemctl stop tomcat
 
 echo "#####  Create new database with a new name  #####"
 
-sudo -u postgres psql <<EOF
+cat <<EOF | psql -U $DSPACE_DATABASE_OWNER -h $DATABASE_HOSTNAME -d postgres
 
 DROP DATABASE IF EXISTS $NEW_DSPACE_DATABASE_NAME;
 
@@ -49,11 +92,11 @@ EOF
 
 echo "#####  Import metadata into the new database  #####"
 
-sudo -u postgres psql -U $DSPACE_DATABASE_OWNER -h localhost $NEW_DSPACE_DATABASE_NAME < $PATH_TO_OLD_DSPACE_DATABASE_DATA
+psql -U $DSPACE_DATABASE_OWNER -h $DATABASE_HOSTNAME $NEW_DSPACE_DATABASE_NAME < $PATH_TO_OLD_DSPACE_DATABASE_DATA
 
 echo "##### Modify the databases #####"
 
-sudo -u postgres psql <<EOF
+cat <<EOF | psql -U $DSPACE_DATABASE_OWNER -h $DATABASE_HOSTNAME -d postgres
 
 ALTER DATABASE $EXISTING_DSPACE_DATABASE_NAME RENAME TO $EXISTING_DSPACE_DATABASE_NAME_CHANGED;
 
@@ -65,9 +108,17 @@ ALTER DATABASE $NEW_DSPACE_DATABASE_NAME RENAME TO $EXISTING_DSPACE_DATABASE_NAM
 
 EOF
 
+echo "#####  Show database info  #####"
+
+sudo /srv/oulib/dspace/bin/dspace database info
+
+echo "#####  Running database upgrade script  #####"
+
+sudo /srv/oulib/dspace/bin/dspace database migrate
+
 echo "#####  Copy over the registry files  #####"
 sudo mv /srv/oulib/dspace/config/registries/ /srv/oulib/dspace/config/registries-old/
-sudo tar -xvf $PATH_TO_REGISTRIES_FROM_OLD_DSPACE -C /tmp
+sudo mkdir /tmp/registries && sudo tar -xvf $PATH_TO_REGISTRIES_FROM_OLD_DSPACE -C /tmp/registries --strip-components=2
 sudo mv /tmp/registries /srv/oulib/dspace/config
 sudo chown -R tomcat:tomcat /srv/oulib/dspace/config/registries/
